+++ util/tasks/godot/threaded_task_gd.h
@@ -1,0 +1,39 @@
+#ifndef ZN_THREADED_TASK_GD
+#define ZN_THREADED_TASK_GD
+
+#include "../threaded_task.h"
+#include <core/object/ref_counted.h>
+#include <core/object/script_language.h> // needed for GDVIRTUAL macro
+#include <core/object/gdvirtual.gen.inc> // Also needed for GDVIRTUAL macro...
+
+namespace zylann {
+
+class ZN_ThreadedTaskInternal;
+
+class ZN_ThreadedTask : public RefCounted {
+	GDCLASS(ZN_ThreadedTask, RefCounted)
+public:
+	void run(int thread_index);
+	int get_priority();
+	bool is_cancelled();
+
+	// Internal
+	bool is_scheduled() const;
+	void mark_completed();
+	IThreadedTask *create_task();
+
+private:
+	GDVIRTUAL1(_run, int);
+	GDVIRTUAL0R(int, _get_priority);
+	GDVIRTUAL0R(bool, _is_cancelled);
+
+	static void _bind_methods();
+
+	// Created upon scheduling, owned by the task runner
+	ZN_ThreadedTaskInternal *_scheduled_task = nullptr;
+	bool _completed = false;
+};
+
+} // namespace zylann
+
+#endif // ZN_THREADED_TASK_GD